requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Let pip choose the best available versions of torch and torchvision for your system
torch
torchvision

numpy==1.26.4
Pillow==10.4.0
scikit-learn==1.5.1
pyyaml==6.0.2
pydantic==2.9.0
python-multipart==0.0.9
streamlit==1.39.0
opencv-python==4.10.0.84



